refactor: folder structure of the architecture
Moved service and repository folder outside as they are constantly being used for other purposes than api
