+++ api/controllers/jwt_auth_controller.go
@@ -1,8 +1,8 @@
 package controllers
 
 import (
-	"github.com/dipeshdulal/clean-gin/api/services"
 	"github.com/dipeshdulal/clean-gin/lib"
+	"github.com/dipeshdulal/clean-gin/services"
 	"github.com/gin-gonic/gin"
 )
